+++ lib/dnsimple/struct/domain.rb
@@ -4,6 +4,13 @@ module Dnsimple
   module Struct
 
     class Domain < Base
+
+      def initialize(attributes = {})
+        attributes.delete("expires_on")
+        super
+        @expires_on = Date.parse(expires_at).to_s if expires_at
+      end
+
       # @return [Integer] The domain ID in DNSimple.
       attr_accessor :id
 
@@ -28,15 +35,27 @@ module Dnsimple
       # @return [Bool] True if the domain WHOIS privacy is enabled, false otherwise.
       attr_accessor :private_whois
 
-      # @return [String] The date the domain will expire.
-      attr_accessor :expires_on
+      # @return [String] The timestamp when domain will expire.
+      attr_accessor :expires_at
 
       # @return [String] When the domain was created in DNSimple.
       attr_accessor :created_at
 
       # @return [String] When the domain was last updated in DNSimple.
       attr_accessor :updated_at
-    end
 
+      # @deprecated Please use #expires_at instead.
+      # @return [String] The date the domain will expire.
+      def expires_on
+        warn "[DEPRECATION] Domain#expires_on is deprecated. Please use `expires_at` instead."
+        @expires_on
+      end
+
+      def expires_on=(expiration_date)
+        warn "[DEPRECATION] Domain#expires_on= is deprecated. Please use `expires_at=` instead."
+        @expires_on = expiration_date
+      end
+
+    end
   end
 end