+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 ## master
 
-- CHANGED: `Domain` now have `expires_at` (timestamp) instead of `expires_on` (date only). (dnsimple/dnsimple-ruby#186)
+- CHANGED: `Domain#expires_on` (date only) is deprecated in favor of `Domain#expires_at` (timestamp). (dnsimple/dnsimple-ruby#186)
 
 ## 5.0.0